re arranged some functions and fixed a bug
even when num_tt_eat was not set by user, we still checked its value.
now we only check if it's set.
