+++ philosophers_working/srcs/main_thread/threads.c
@@ -6,7 +6,7 @@
 /*   By: dpentlan <dpentlan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/16 16:18:43 by dpentlan          #+#    #+#             */
-/*   Updated: 2023/10/19 12:28:31 by dpentlan         ###   ########.fr       */
+/*   Updated: 2023/10/19 12:49:56 by dpentlan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,7 +32,8 @@ static bool	ft_create_philos(t_info *info)
 		info->philos[i].id = i;
 		info->philos[i].info = info;
 		info->philos[i].last_ate = info->start_time;
-		info->philos[i].num_tt_eat = info->num_tt_eat;
+		if (info->num_tt_eat)
+			info->philos[i].num_tt_eat = info->num_tt_eat;
 		if (pthread_create(&info->philos[i].thread_id, NULL,
 				ft_pthread_entry_point, &info->philos[i]))
 			return (printf("Error: Could not create all philosophers\n"), 1);
